+++ lib/src/confetti.dart
@@ -1,5 +1,7 @@
 import 'dart:math';
 
+import 'package:confetti/src/particle_stats.dart';
+import 'package:confetti/src/constants.dart';
 import 'package:confetti/src/particle.dart';
 import 'package:flutter/material.dart';
 
@@ -26,8 +28,9 @@ class ConfettiWidget extends StatefulWidget {
     this.maximumSize = const Size(30, 15),
     this.particleDrag = 0.05,
     this.canvas,
-    this.child,
+    this.pauseEmissionOnLowFrameRate = true,
     this.createParticlePath,
+    this.child,
   })  : assert(
           emissionFrequency >= 0 &&
               emissionFrequency <= 1 &&
@@ -54,10 +57,10 @@ class ConfettiWidget extends StatefulWidget {
   /// life. The default [minBlastForce] is set to `5`
   final double minBlastForce;
 
-  /// The [blastDirectionality] is an enum that takes one of two
-  /// values - directional or explosive.
+  /// {@macro blast_directionality}
   ///
-  /// The default is set to directional
+  /// The default value is [BlastDirectionality.directional], the direction
+  /// can be set with [blastDirection].
   final BlastDirectionality blastDirectionality;
 
   /// The [blastDirection] is a radial value to determine the direction of the
@@ -67,16 +70,17 @@ class ConfettiWidget extends StatefulWidget {
   /// A value of `PI` will emit to the left of the canvas/screen.
   final double blastDirection;
 
-  /// The [createParticlePath] is optional function that returns custom Path
-  /// needed to generate particles.
+  /// The [createParticlePath] is an optional function that returns a custom
+  /// `Path` to generate particles.
   ///
-  /// The default function returns rectangular path
+  /// The default function returns a rectangular path.
   final Path Function(Size size)? createParticlePath;
 
   /// The [gravity] is the speed at which the confetti will fall.
   /// The higher the [gravity] the faster it will fall.
   ///
   /// It can be set to a value between `0` and `1`
+  ///
   /// Default value is `0.1`
   final double gravity;
 
@@ -129,8 +133,13 @@ class ConfettiWidget extends StatefulWidget {
   /// An optional parameter to specify the area size where the confetti will
   /// be thrown.
   ///
-  /// By default this is set to then screen size.
+  /// By default this is set to the window size.
   final Size? canvas;
+
+  /// If `true` new particles will not be created if the FPS is lower
+  /// than 60. Default is `true`, set to `false` to ensure particles are always
+  /// created, regardless of frame rate.
+  final bool pauseEmissionOnLowFrameRate;
 
   /// Child widget to display
   final Widget? child;
@@ -148,10 +157,11 @@ class _ConfettiWidgetState extends State<ConfettiWidget>
   late ParticleSystem _particleSystem;
 
   /// Keeps track of emition position on screen layout changes
-  Offset? _emitterPosition;
+  late Offset _emitterPosition;
 
-  /// Keeps track of the screen size on layout changes
-  /// Controls the sizing restrictions for when confetti should be vissible
+  /// Keeps track of the screen size on layout changes.
+  ///
+  /// Controls the sizing restrictions for when confetti should be visible.
   Size _screenSize = const Size(0, 0);
 
   @override
@@ -160,18 +170,19 @@ class _ConfettiWidgetState extends State<ConfettiWidget>
     widget.confettiController.addListener(_handleChange);
 
     _particleSystem = ParticleSystem(
-        emissionFrequency: widget.emissionFrequency,
-        numberOfParticles: widget.numberOfParticles,
-        maxBlastForce: widget.maxBlastForce,
-        minBlastForce: widget.minBlastForce,
-        gravity: widget.gravity,
-        blastDirection: widget.blastDirection,
-        blastDirectionality: widget.blastDirectionality,
-        colors: widget.colors,
-        minimumSize: widget.minimumSize,
-        maximumSize: widget.maximumSize,
-        particleDrag: widget.particleDrag,
-        createParticlePath: widget.createParticlePath);
+      emissionFrequency: widget.emissionFrequency,
+      numberOfParticles: widget.numberOfParticles,
+      maxBlastForce: widget.maxBlastForce,
+      minBlastForce: widget.minBlastForce,
+      gravity: widget.gravity,
+      blastDirection: widget.blastDirection,
+      blastDirectionality: widget.blastDirectionality,
+      colors: widget.colors,
+      minimumSize: widget.minimumSize,
+      maximumSize: widget.maximumSize,
+      particleDrag: widget.particleDrag,
+      createParticlePath: widget.createParticlePath,
+    );
 
     _particleSystem.addListener(_particleSystemListener);
 
@@ -186,10 +197,12 @@ class _ConfettiWidgetState extends State<ConfettiWidget>
       ..addListener(_animationListener)
       ..addStatusListener(_animationStatusListener);
 
-    if (widget.confettiController.state == ConfettiControllerState.playing) {
-      _startAnimation();
-      _startEmission();
-    }
+    WidgetsBinding.instance.addPostFrameCallback((_) {
+      if (widget.confettiController.state == ConfettiControllerState.playing) {
+        _startAnimation();
+        _startEmission();
+      }
+    });
   }
 
   void _handleChange() {
@@ -199,15 +212,40 @@ class _ConfettiWidgetState extends State<ConfettiWidget>
     } else if (widget.confettiController.state ==
         ConfettiControllerState.stopped) {
       _stopEmission();
+    } else if (widget.confettiController.state ==
+        ConfettiControllerState.stoppedAndCleared) {
+      _stopEmission(clearAllParticles: true);
+    } else if (widget.confettiController.state ==
+        ConfettiControllerState.disposed) {
+      _stopEmission(clearAllParticles: true);
     }
   }
+
+  late var lastTime = DateTime.now().millisecondsSinceEpoch;
 
   void _animationListener() {
     if (_particleSystem.particleSystemStatus == ParticleSystemStatus.finished) {
       _animController.stop();
       return;
     }
-    _particleSystem.update();
+    final currentTime = DateTime.now().millisecondsSinceEpoch;
+    final deltaTime = (currentTime - lastTime) / 1000;
+
+    lastTime = currentTime;
+
+    if (deltaTime > kLowLimit) {
+      _particleSystem.update(kLowLimit,
+          pauseEmission: widget.pauseEmissionOnLowFrameRate);
+    } else {
+      _particleSystem.update(deltaTime);
+    }
+
+    widget.confettiController.particleStatsCallback?.call(
+      ParticleStats(
+        numberOfParticles: _particleSystem.numberOfParticles,
+        activeNumberOfParticles: _particleSystem.activeNumberOfParticles,
+      ),
+    );
   }
 
   void _animationStatusListener(AnimationStatus status) {
@@ -229,22 +267,20 @@ class _ConfettiWidgetState extends State<ConfettiWidget>
     _particleSystem.startParticleEmission();
   }
 
-  void _stopEmission() {
+  void _stopEmission({bool clearAllParticles = false}) {
     if (_particleSystem.particleSystemStatus == ParticleSystemStatus.stopped) {
       return;
     }
-    _particleSystem.stopParticleEmission();
+    _particleSystem.stopParticleEmission(clearAllParticles: clearAllParticles);
   }
 
   void _startAnimation() {
     // Make sure widgets are built before setting screen size and position
-    WidgetsBinding.instance.addPostFrameCallback((_) {
-      if (mounted) {
-        _setScreenSize();
-        _setEmitterPosition();
-        _animController.forward(from: 0);
-      }
-    });
+    if (mounted) {
+      _setScreenSize();
+      _setEmitterPosition();
+      _animController.forward(from: 0);
+    }
   }
 
   void _stopAnimation() {
@@ -267,13 +303,28 @@ class _ConfettiWidgetState extends State<ConfettiWidget>
   }
 
   Offset _getContainerPosition() {
-    final containerRenderBox =
-        _particleSystemKey.currentContext!.findRenderObject() as RenderBox;
-    return containerRenderBox.localToGlobal(Offset.zero);
+    if (mounted) {
+      final containerRenderBox =
+          _particleSystemKey.currentContext?.findRenderObject() as RenderBox?;
+      return containerRenderBox?.localToGlobal(Offset.zero) ?? Offset.zero;
+    } else {
+      return Offset.zero;
+    }
   }
 
   Size _getScreenSize() {
-    return widget.canvas ?? MediaQuery.of(context).size;
+    if (mounted) {
+      try {
+        return widget.canvas ?? MediaQuery.of(context).size;
+      } catch (e) {
+        // Ugh flutter debug on web throws an error here. Need to clean this
+        // whole thing up.
+        print('[flutter_confetti] Error getting screen size: $e');
+        return widget.canvas ?? Size.zero;
+      }
+    } else {
+      return widget.canvas ?? Size.zero;
+    }
   }
 
   /// On layout change update the position of the emitter
@@ -285,29 +336,37 @@ class _ConfettiWidgetState extends State<ConfettiWidget>
   /// The emitter position is first set in the `addPostFrameCallback`
   /// in [initState].
   void _updatePositionAndSize() {
+    // TODO: improve this
     if (_getScreenSize() != _screenSize) {
       _setScreenSize();
-      if (_emitterPosition != null) {
-        _setEmitterPosition();
-      }
+      _setEmitterPosition();
     }
   }
 
   @override
   Widget build(BuildContext context) {
-    _updatePositionAndSize();
-    return RepaintBoundary(
-      child: CustomPaint(
-        key: _particleSystemKey,
-        foregroundPainter: ParticlePainter(
-          _animController,
-          strokeWidth: widget.strokeWidth,
-          strokeColor: widget.strokeColor,
-          particles: _particleSystem.particles,
-          paintEmitterTarget: widget.displayTarget,
-        ),
-        child: widget.child,
-      ),
+    return LayoutBuilder(
+      builder: (BuildContext context, BoxConstraints constraints) {
+        // Update position and size only when layout changes
+        WidgetsBinding.instance.addPostFrameCallback((_) {
+          _updatePositionAndSize();
+        });
+
+        return RepaintBoundary(
+          child: CustomPaint(
+            key: _particleSystemKey,
+            willChange: true,
+            foregroundPainter: ParticlePainter(
+              _animController,
+              strokeWidth: widget.strokeWidth,
+              strokeColor: widget.strokeColor,
+              particles: _particleSystem.particles,
+              paintEmitterTarget: widget.displayTarget,
+            ),
+            child: widget.child,
+          ),
+        );
+      },
     );
   }
 
@@ -373,11 +432,45 @@ class ParticlePainter extends CustomPainter {
 
   void _paintParticles(Canvas canvas) {
     for (final particle in particles) {
-      final rotationMatrix4 = Matrix4.identity()
-        ..translate(particle.location.dx, particle.location.dy)
-        ..rotateX(particle.angleX)
-        ..rotateY(particle.angleY)
-        ..rotateZ(particle.angleZ);
+      if (!particle.active) continue;
+
+      // OG way to do it.
+      // final rotationMatrix4 = Matrix4.identity()
+      //   ..setEntry(3, 2, 0.001) // perspective
+      //   ..translate(particle.location.dx, particle.location.dy)
+      //   ..rotateX(particle.angleX)
+      //   ..rotateY(particle.angleY)
+      //   ..rotateZ(particle.rotateZ ? particle.angleZ : 0.0);
+
+      // Manual way to do it. Should be more performant, and less memory usage.
+      // Should be the same as above... Looking good to me.
+      final dx = particle.location.dx;
+      final dy = particle.location.dy;
+      final cosX = cos(particle.angleX);
+      final sinX = sin(particle.angleX);
+      final cosY = cos(particle.angleY);
+      final sinY = sin(particle.angleY);
+      final cosZ = particle.rotateZ ? cos(particle.angleZ) : 1.0;
+      final sinZ = particle.rotateZ ? sin(particle.angleZ) : 0.0;
+
+      final rotationMatrix4 = Matrix4(
+        cosY * cosZ,
+        cosX * sinZ + sinX * sinY * cosZ,
+        sinX * sinZ - cosX * sinY * cosZ,
+        0.0,
+        -cosY * sinZ,
+        cosX * cosZ - sinX * sinY * sinZ,
+        sinX * cosZ + cosX * sinY * sinZ,
+        0.0,
+        sinY,
+        -sinX * cosY,
+        cosX * cosY,
+        0.001, // perspective
+        dx,
+        dy,
+        0.0,
+        1.0,
+      );
 
       final finalPath = particle.path.transform(rotationMatrix4.storage);
       canvas.drawPath(finalPath, _particlePaint..color = particle.color);
@@ -393,23 +486,48 @@ class ParticlePainter extends CustomPainter {
   }
 }
 
+/// {@template particle_stats_callback}
+/// This callback provides [ParticleStats] as an argument.
+/// {@endtemplate}
+typedef ParticleStatsCallback = void Function(ParticleStats stats);
+
 class ConfettiController extends ChangeNotifier {
-  ConfettiController({this.duration = const Duration(seconds: 30)})
-      : assert(!duration.isNegative && duration.inMicroseconds > 0);
+  ConfettiController({
+    this.duration = const Duration(seconds: 30),
+    this.particleStatsCallback,
+  }) : assert(!duration.isNegative && duration.inMicroseconds > 0);
 
   Duration duration;
 
   ConfettiControllerState _state = ConfettiControllerState.stopped;
 
+  /// {@macro confetti_controller_state}
   ConfettiControllerState get state => _state;
+
+  /// {@macro particle_stats_callback}
+  final ParticleStatsCallback? particleStatsCallback;
 
   void play() {
     _state = ConfettiControllerState.playing;
     notifyListeners();
   }
 
-  void stop() {
-    _state = ConfettiControllerState.stopped;
+  void stop({bool clearAllParticles = false}) {
+    // if state is already disposed, it can not be stopped.
+    if (_state == ConfettiControllerState.disposed) return;
+
+    if (clearAllParticles) {
+      _state = ConfettiControllerState.stoppedAndCleared;
+    } else {
+      _state = ConfettiControllerState.stopped;
+    }
     notifyListeners();
+  }
+
+  @override
+  void dispose() {
+    _state = ConfettiControllerState.disposed;
+    notifyListeners();
+    super.dispose();
   }
 }