+++ lib/src/constants.dart
@@ -1,0 +1,2 @@
+const double kLowLimit = 1 / 60;
+const desiredSpeed = 1 / kLowLimit;